refactor(output): extract config tests to sibling file
